move checkpoint_seeder to cpu_node
